% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_assess_plausibility.R
\name{print.plausibility_assessment}
\alias{print.plausibility_assessment}
\title{Print method for plausibility assessment}
\usage{
\method{print}{plausibility_assessment}(x, ...)
}
\arguments{
\item{x}{An object of class "plausibility_assessment"}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print method for plausibility assessment
}
\keyword{internal}
