% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tipse-package.R
\docType{package}
\name{tipse-package}
\alias{tipse}
\alias{tipse-package}
\title{tipse: Tipping Point Analysis for Survival Endpoints}
\description{
Implements tipping point sensitivity analysis for time-to-event endpoints under different missing data scenarios, as described in Oodally et al. (2025) \doi{10.48550/arXiv.2506.19988}. Supports both model-based and model-free imputation, multiple imputation workflows, plausibility assessment and visualizations. Enables robust assessment for regulatory and exploratory analyses.
}
\author{
\strong{Maintainer}: Ajmal Oodally \email{ajmal.oodally@novartis.com} (\href{https://orcid.org/0000-0001-8143-5011}{ORCID})

Authors:
\itemize{
  \item Craig Wang \email{craig.wang@novartis.com} (\href{https://orcid.org/0000-0003-1804-2463}{ORCID})
}

Other contributors:
\itemize{
  \item Zheng Li \email{zheng.li@novartis.com} (\href{https://orcid.org/0000-0002-7655-0290}{ORCID}) [contributor]
}

}
\keyword{internal}
